Moved the dateStrings code out to datetime_decode.js.
Added testing for the dateStrings code.
