+++ index.js
@@ -39,7 +39,6 @@ function ZongJi(dsn, options) {
 
     binlogDsn = dsn;
   }
-  this.connectionConfig = cloneObjectSimple(binlogDsn);
   this.ctrlCallbacks = [];
 
   this.connection = mysql.createConnection(binlogDsn);
