+++ lib/common.js
@@ -1,5 +1,6 @@
 var iconv = require('iconv-lite');
 var decodeJson = require('./json_decode');
+var dtDecode = require('./datetime_decode');
 
 var MysqlTypes = exports.MysqlTypes = {
   DECIMAL: 0,
@@ -351,114 +352,6 @@ var readTemporalFraction = function(parser, fractionPrecision) {
   };
 };
 
-// fraction is the fractional second object from readTemporalFraction().
-// returns '' or a '.' followed by fraction.precision digits, like '.123'
-var getFractionString = function(fraction) {
-  return fraction ? '.' + zeroPad(fraction.value, fraction.precision) : '';
-};
-
-// 1950-00-00 and the like are perfectly valid Mysql dateStrings.  A 0 portion
-// of a date is essentially a null part of the date, so we should keep it.
-// year, month, and date must be integers >= 0.  January is month === 1.
-var getDateString = function(year, month, date){
-  return zeroPad(year, 4) + '-' +
-         zeroPad(month, 2) + '-' +
-         zeroPad(date, 2);
-};
-
-// Date object months are 1 less than Mysql months, and we need to filter 0.
-// If we don't filter 0, 2017-00-01 will become the javascript Date 2016-12-01,
-// which is not what it means.  It means 2017-NULL-01, but the Date object
-// cannot handle it, so we want to return an invalid month, rather than a
-// subtracted month.
-var jsMonthFromMysqlMonth = function(month){
-  return month > 0 ? month - 1 : undefined;
-};
-
-// dateStrings are used only if the dateStrings option is true, or is an array
-// containing the sql type name string, 'DATE', 'DATETIME', or 'TIMESTAMP'.
-// This follows the documentation of the dateStrings option in node-mysql.
-var useDateStringsForType = function(dateStrings, sqlTypeString){
-  return dateStrings &&
-           (dateStrings === true ||
-            dateStrings.indexOf && dateStrings.indexOf(sqlTypeString) > -1);
-};
-
-// Returns a new Date object or Mysql dateString, following the dateStrings
-// option.  With the dateStrings option, it can output valid Mysql DATE strings
-// representing values that cannot be represented by a Date object, such as
-// values with a null part like '1950-00-04', or a zero date '0000-00-00'.
-var getDate = function(dateStrings, // node-mysql dateStrings option
-                       year,
-                       month, // January === 1
-                       date
-                      )
-{
-  if(!useDateStringsForType(dateStrings, 'DATE')){
-    return new Date(year,
-                    jsMonthFromMysqlMonth(month),
-                    date);
-  }
-  return getDateString(year, month, date);
-};
-
-// Returns a new Date object or Mysql dateString, following the dateStrings
-// option.  Fraction is an optional parameter that comes from
-// readTemporalFraction().  Mysql dateStrings are needed for microsecond
-// precision, or to represent '0000-00-00 00:00:00'.
-var getDateTime = function(dateStrings, // node-mysql dateStrings option
-                           year,
-                           month,       // January === 1
-                           date,
-                           hour,
-                           minute,
-                           second,
-                           fraction     // optional fraction of second object
-                          )
-{
-  if(!useDateStringsForType(dateStrings, 'DATETIME')){
-    return new Date(year,
-                    jsMonthFromMysqlMonth(month),
-                    date,
-                    hour,
-                    minute,
-                    second,
-                    fraction ? fraction.milliseconds : 0);
-  }
-  return getDateString(year, month, date) + ' ' +
-         zeroPad(hour, 2) + ':' +
-         zeroPad(minute, 2) + ':' +
-         zeroPad(second, 2) +
-         getFractionString(fraction);
-};
-
-// Returns a new Date object or Mysql dateString, following the dateStrings
-// option.  Fraction is an optional parameter that comes from
-// readTemporalFraction().  With the dateStrings option from node-mysql,
-// this returns a mysql TIMESTAMP string, like '1975-03-01 23:03:20.38945' or
-// '1975-03-01 00:03:20'.  Mysql strings are needed for precision beyond ms.
-var getTimeStamp = function(dateStrings, // node-mysql dateStrings option
-                            secondsFromEpoch, // an integer
-                            fraction // optional fraction of second object
-                           )
-{
-  var milliseconds = fraction ? fraction.milliseconds : 0;
-  var dateObject = new Date(secondsFromEpoch * 1000 + milliseconds);
-  if(!useDateStringsForType(dateStrings, 'TIMESTAMP')){
-    return dateObject;
-  }
-  if(secondsFromEpoch === 0 && (!fraction || fraction.value === 0)){
-    return '0000-00-00 00:00:00' + getFractionString(fraction);
-  }
-  return getDateString(dateObject.getFullYear(),
-                       dateObject.getMonth() + 1,
-                       dateObject.getDate()) + ' ' +
-         zeroPad(dateObject.getHours(), 2) + ':' +
-         zeroPad(dateObject.getMinutes(), 2) + ':' +
-         zeroPad(dateObject.getSeconds(), 2) +
-         getFractionString(fraction);
-};
-
 // This function is used to read and interpret non-null values from parser.
 // The parser object contains the raw value, and functions useful for value
 // interpretation, but not the underlying column type or how the value ought to
@@ -572,8 +465,8 @@ exports.readMysqlValue = function(
       break;
     case MysqlTypes.DATE:
       var raw = exports.parseUInt24(parser);
-      result = getDate(
-        zongji.connectionConfig.dateStrings, // node-mysql dateStrings option
+      result = dtDecode.getDate(
+        zongji.connection.config.dateStrings, // node-mysql dateStrings option
         sliceBits(raw, 9, 24),     // year
         sliceBits(raw, 5, 9),      // month
         sliceBits(raw, 0, 5)       // day
@@ -617,15 +510,15 @@ exports.readMysqlValue = function(
                zeroPad(second, 2);
 
       if(fraction !== false){
-        result += getFractionString(fraction);
+        result += dtDecode.getFractionString(fraction);
       }
       break;
     case MysqlTypes.DATETIME:
       var raw = exports.parseUInt64(parser);
       var date = Math.floor(raw / 1000000);
       var time = raw % 1000000;
-      result = getDateTime(
-        zongji.connectionConfig.dateStrings,  // node-mysql dateStrings option
+      result = dtDecode.getDateTime(
+        zongji.connection.config.dateStrings,  // node-mysql dateStrings option
         Math.floor(date / 10000),             // year
         Math.floor((date % 10000) / 100),     // month
         date % 100,                           // day
@@ -642,8 +535,8 @@ exports.readMysqlValue = function(
       var fraction = readTemporalFraction(parser, column.metadata.decimals);
 
       var yearMonth = sliceBits(rawHigh, 14, 31);
-      result = getDateTime(
-        zongji.connectionConfig.dateStrings, // node-mysql dateStrings option
+      result = dtDecode.getDateTime(
+        zongji.connection.config.dateStrings, // node-mysql dateStrings option
         Math.floor(yearMonth / 13), // year
         (yearMonth % 13),           // month
         sliceBits(rawLow, 17, 22),  // day
@@ -655,15 +548,15 @@ exports.readMysqlValue = function(
       break;
     case MysqlTypes.TIMESTAMP:
       var raw = parser.parseUnsignedNumber(4);
-      result = getTimeStamp(zongji.connectionConfig.dateStrings, raw);
+      result = dtDecode.getTimeStamp(zongji.connection.config.dateStrings, raw);
       break;
     case MysqlTypes.TIMESTAMP2:
       var raw = readIntBE(parser._buffer, parser._offset, 4);
       parser._offset += 4;
       var fraction = readTemporalFraction(parser, column.metadata.decimals);
-      result = getTimeStamp(zongji.connectionConfig.dateStrings,
-                            raw,       // seconds from epoch
-                            fraction); // fraction of a second object
+      result = dtDecode.getTimeStamp(zongji.connection.config.dateStrings,
+                                     raw,       // seconds from epoch
+                                     fraction); // fraction of a second object
       break;
     case MysqlTypes.YEAR:
       var raw = parser.parseUnsignedNumber(1);
